// frontend/src/components/Checkout/ProductInfo/types.ts
export interface AddressFormData {
    fullName: string;
    phone: string;
    email: string;
    city: string;
    cityCode: any;
    districtCode: any;
    district: any;
    wardCode: any;
    ward: any;
    address: any;
    notes?: string;
  }
  
  export interface OrdererInfo {
    fullName: string;
    phone: string;
    email: string;
  }
  export interface OrdererInfoPickup {
    fullName: string;
    phone: string;
    email: string;
  }
  